Restructure for packaging
Working--
