--- geopolrisk/requirements.txt
@@ -1,3 +1,0 @@
-python==3.11.5
-pandas==2.0.3
-tqdm==4.65.0
