--- geopolrisk/tests/defaults.py
@@ -1,199 +1,0 @@
-# Copyright (C) 2023 University of Bordeaux, CyVi Group & Anish Koyamparambath
-# This file is part of geopolrisk-py library.
-#
-# geopolrisk-py is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# geopolrisk-py is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with geopolrisk-py.  If not, see <https://www.gnu.org/licenses/>.
-
-ListofMetals = [
-    250410,
-    251110,
-    251910,
-    252010,
-    252400,
-    252910,
-    252922,
-    260111,
-    260200,
-    260300,
-    260400,
-    260600,
-    260700,
-    260800,
-    260900,
-    261000,
-    261100,
-    261210,
-    261310,
-    261400,
-    261510,
-    261590,
-    261610,
-    261710,
-    270111,
-    270112,
-    270900,
-    271111,
-    280450,
-    280480,
-    280490,
-    280540,
-    282200,
-    282560,
-    283691,
-    284690,
-    710811,
-    711011,
-    711021,
-    711031,
-    810710,
-    811291,
-]
-
-
-ListofYears = list(range(2018, 2023))
-
-ListofCountries = [
-    36,
-    40,
-    56,
-    124,
-    152,
-    156,
-    170,
-    188,
-    208,
-    233,
-    246,
-    251,
-    276,
-    300,
-    352,
-    372,
-    376,
-    380,
-    392,
-    410,
-    428,
-    440,
-    442,
-    484,
-    528,
-    554,
-    579,
-    616,
-    620,
-    703,
-    705,
-    724,
-    752,
-    757,
-    792,
-    826,
-    842,
-    348,
-    203,
-]
-
-ListofCountryName = [
-    "Australia",
-    "Austria",
-    "Belgium",
-    "Canada",
-    "Chile",
-    "China",
-    "Colombia",
-    "Costa Rica",
-    "Denmark",
-    "Estonia",
-    "Finland",
-    "France",
-    "Germany",
-    "Greece",
-    "Iceland",
-    "Ireland",
-    "Israel",
-    "Italy",
-    "Japan",
-    "Rep. of Korea",
-    "Latvia",
-    "Lithuania",
-    "Luxembourg",
-    "Mexico",
-    "Netherlands",
-    "New Zealand",
-    "Norway",
-    "Poland",
-    "Portugal",
-    "Slovakia",
-    "Slovenia",
-    "Spain",
-    "Sweden",
-    "Switzerland",
-    "Turkey",
-    "United Kingdom",
-    "USA",
-    "Hungary",
-    "Czechia",
-]
-
-ListofMetalName = [
-    "Graphite",
-    "Baryte",
-    "Magnesium",
-    "Gypsum",
-    "Asbestos",
-    "Feldspar",
-    "Fluorspar",
-    "Iron",
-    "Manganese",
-    "Copper",
-    "Nickel",
-    "Aluminium",
-    "Lead",
-    "Zinc",
-    "Tin",
-    "Chromium",
-    "Tungsten",
-    "Uranium",
-    "Molybdenum",
-    "Titanium",
-    "Zirconium",
-    "Niobium",
-    "Silver",
-    "Antimony",
-    "Coal",
-    "Lignite",
-    "Crude oil",
-    "Natural gas",
-    "Boron",
-    "Arsenic",
-    "Selenium",
-    "Mercury",
-    "Cobalt",
-    "Germanium",
-    "Lithium",
-    "Rare earth",
-    "Gold",
-    "Platinum",
-    "Palladium",
-    "Rhodium",
-    "Cadmium",
-    "Gallium",
-]
-
-
-import random
-
-
-def random_sampler(A, n):
-    return random.sample(A, n)
