+++ src/geopolrisk/assessment/core.py
