+++ src/geopolrisk/lib/wgi.db
