+++ src/geopolrisk/lib/world_mining_data.db
